refactor (directory) renaming directories and making menu bar small
