+++ CycleCare/app/src/main/AndroidManifest.xml
@@ -16,16 +16,16 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".activities.viewcontentbymedic.ViewContentByMedic"
+            android:name=".activities.view_content_medic_pov.ViewContentByMedic"
             android:exported="false" />
         <activity
-            android:name=".activities.viewcontentbymedic.MyContentActivity"
+            android:name=".activities.view_content_medic_pov.MyContentActivity"
             android:exported="false"/>
         <activity
-            android:name=".activities.registerreminder.RegisterReminder"
+            android:name=".activities.register_reminder.RegisterReminder"
             android:exported="false" />
         <activity
-            android:name=".activities.principalscreen.PrincipalScreen"
+            android:name=".activities.main_screen.PrincipalScreen"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -34,22 +34,22 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activities.viewcontentbymedic.PublishInformativeContentActivity"
+            android:name=".activities.view_content_medic_pov.PublishInformativeContentActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.registeraccount.PeriodInformationActivity"
+            android:name=".activities.register_account.PeriodInformationActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.registeraccount.CycleTypeActivity"
+            android:name=".activities.register_account.CycleTypeActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.registeraccount.RegisterAccountActivity"
+            android:name=".activities.register_account.RegisterAccountActivity"
             android:exported="false"/>
         <activity
-            android:name=".activities.verifyemail.VerifyEmailActivity"
+            android:name=".activities.verify_email.VerifyEmailActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.forgotpassword.ForgotPasswordActivity"
+            android:name=".activities.forgot_password.ForgotPasswordActivity"
             android:exported="false" />
         <activity
             android:name=".activities.login.LoginActivity"
