+++ CycleCare/app/src/main/java/com/ikariscraft/cyclecare/activities/forgot_password/ForgotPasswordActivity.java
@@ -1,4 +1,4 @@
-package com.ikariscraft.cyclecare.activities.verifyemail;
+package com.ikariscraft.cyclecare.activities.forgot_password;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.ikariscraft.cyclecare.R;
 
-public class VerifyEmailActivity extends AppCompatActivity {
+public class ForgotPasswordActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_verify_email);
+        setContentView(R.layout.activity_forgot_password);
     }
 }
