+++ CycleCare/app/src/main/java/com/ikariscraft/cyclecare/activities/register_account/CycleTypeActivity.java
@@ -1,4 +1,4 @@
-package com.ikariscraft.cyclecare.activities.registeraccount;
+package com.ikariscraft.cyclecare.activities.register_account;
 
 import androidx.appcompat.app.AppCompatActivity;
