+++ CycleCare/app/src/main/java/com/ikariscraft/cyclecare/activities/register_account/PeriodInformationViewModel.java
@@ -1,4 +1,4 @@
-package com.ikariscraft.cyclecare.activities.registeraccount;
+package com.ikariscraft.cyclecare.activities.register_account;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
