+++ CycleCare/app/src/main/java/com/ikariscraft/cyclecare/activities/verify_email/VerifyEmailActivity.java
@@ -1,4 +1,4 @@
-package com.ikariscraft.cyclecare.activities.viewcontentbymedic;
+package com.ikariscraft.cyclecare.activities.verify_email;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.ikariscraft.cyclecare.R;
 
-public class MyContentActivity extends AppCompatActivity {
+public class VerifyEmailActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my_content);
+        setContentView(R.layout.activity_verify_email);
     }
 }
