+++ CycleCare/app/src/main/java/com/ikariscraft/cyclecare/activities/view_content_medic_pov/MyContentActivity.java
@@ -1,4 +1,4 @@
-package com.ikariscraft.cyclecare.activities.viewcontentbymedic;
+package com.ikariscraft.cyclecare.activities.view_content_medic_pov;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.ikariscraft.cyclecare.R;
 
-public class PublishInformativeContentActivity extends AppCompatActivity {
+public class MyContentActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_publish_informative_content);
+        setContentView(R.layout.activity_my_content);
     }
 }
