+++ CycleCare/app/src/main/res/layout/activity_principal_screen.xml
@@ -20,7 +20,7 @@
             <androidx.fragment.app.FragmentContainerView
                 android:id="@+id/fragment_container"
                 android:layout_width="match_parent"
-                android:layout_height="600dp" />
+                android:layout_height="630dp" />
 
         </RelativeLayout>
