+++ CycleCare/app/src/main/res/layout/fragment_informative_contente.xml
@@ -9,7 +9,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activities.viewcontentfragmentbyuser.InformativeContenteFragment">
+        tools:context=".activities.view_content_user_pov.InformativeContenteFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
