+++ CycleCare/app/src/main/res/layout/fragment_view_content.xml
@@ -16,7 +16,7 @@
         <FrameLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            tools:context=".activities.viewcontentfragmentbyuser.ViewContentFragment">
+            tools:context=".activities.view_content_user_pov.ViewContentFragment">
 
             <LinearLayout
                 android:layout_width="match_parent"
